Round_1B/requirements.txt
PyMuPDF==1.23.26
pdfplumber==0.10.3
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
sentence-transformers==2.2.2
transformers==4.33.3
huggingface_hub<0.16
faiss-cpu==1.7.4
rank-bm25==0.2.2
networkx==3.2.1
nltk==3.8.1
regex==2023.10.3
jsonschema==4.20.0